히스토리를 Round 안에서 관리하도록 변경

+++ RacingGame/src/main/java/com/tdd/racing/entity/impl/DefaultCar.java
@@ -3,7 +3,7 @@ package com.tdd.racing.entity.impl;
 import com.tdd.racing.entity.Car;
 
 public class DefaultCar implements Car {
-    protected double speed; //속도
+    protected int speed; //속도
     private int number;     //차 번호(PK)
 
     public DefaultCar(int number) {
@@ -11,7 +11,7 @@ public class DefaultCar implements Car {
         this.speed = 1;
     }
 
-    public DefaultCar(int number, double speed) {
+    public DefaultCar(int number, int speed) {
         this.number = number;
         this.speed = speed;
     }
@@ -20,7 +20,7 @@ public class DefaultCar implements Car {
         return this.speed;
     }
 
-    public void setSpeed(double speed) {
+    public void setSpeed(int speed) {
         this.speed = speed;
     }
